moved ScrollCatrousel into new folder structure
