+++ app/src/main/kotlin/com/platform/openemoji/MainActivity.kt
@@ -9,8 +9,7 @@ import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Surface
 import androidx.compose.ui.Modifier
 import com.platform.openemoji.screens.SearchScreen
-import com.platform.openemoji.emoji.EmojiIcon
-import com.platform.openemoji.ui.theme.OpenEmojiPlatformTheme
+import com.platform.openemoji.theme.OpenEmojiPlatformTheme
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -23,7 +22,7 @@ class MainActivity : ComponentActivity() {
                 ) {
                     Column {
                         SearchScreen()
-                        EmojiIcon()
+                        // EmojiIcon()
                     }
                 }
             }
