+++ app/src/main/kotlin/com/platform/openemoji/emoji/category/Category.kt
@@ -1,0 +1,3 @@
+package com.platform.openemoji.emoji.category
+
+data class Category(val name: String, val route: String)
